first-commit

+++ src/main/java/com/spring_data_jpa/event_manager/services/HostService.java
@@ -48,8 +48,14 @@ public class HostService {
 
 
     @Transactional
-    public void deleteHost(UUID id){
-        hostRepository.deleteById(id);
+    public boolean deleteHost(UUID id){
+
+        if (hostRepository.existsById(id)) {
+            hostRepository.deleteById(id);
+            return true;
+        }
+        return false;
+
     }
 
 }
